+++ admin/src/main/resources/io/buoyant/admin/js/main-linkerd.js
@@ -28,16 +28,28 @@ require([
   'src/admin',
   'src/dashboard',
   'src/delegate',
+  'src/namerd',
+  'src/logging'
 ], function (
   $, _, bootstrap,
   adminPage,
   dashboard,
-  linkerdDtabPlayground
+  linkerdDtabPlayground,
+  namerd,
+  loggingConfig
 ) {
   // poor man's routing
   if (window.location.pathname.endsWith("/delegator")) {
     adminPage.initialize(true).done(function() {
       new linkerdDtabPlayground();
+    });
+  } else if (window.location.pathname.endsWith("/namerd")) {
+    adminPage.initialize().done(function() {
+      new namerd();
+    });
+  } else if (window.location.pathname.endsWith("/logging")) {
+    adminPage.initialize().done(function() {
+      new loggingConfig();
     });
   } else {
     adminPage.initialize().done(function() {