+++ project/LinkerdBuild.scala
@@ -8,6 +8,7 @@ import scoverage.ScoverageKeys._
 
 object LinkerdBuild extends Base {
   import Base._
+  import Grpc._
 
   val Bundle = config("bundle")
   val Dcos = config("dcos") extend Bundle
@@ -429,8 +430,9 @@ object LinkerdBuild extends Base {
         .dependsOn(core, Router.mux)
 
       val thrift = projectDir("linkerd/protocol/thrift")
-        .dependsOn(core, Router.thrift)
+        .dependsOn(core, Router.thrift % "compile->compile;test->test;e2e->e2e")
         .withTests()
+        .withE2e()
 
       val benchmark = projectDir("linkerd/protocol/benchmark")
         .dependsOn(http, testUtil)
@@ -454,8 +456,9 @@ object LinkerdBuild extends Base {
       val usage = projectDir("linkerd/telemeter/usage")
         .dependsOn(core % "compile->compile;test->test")
         .dependsOn(Namer.core  % "compile->compile;test->test")
-        .dependsOn(Protocol.http, Grpc.usage)
+        .dependsOn(Protocol.http)
         .withLibs(Deps.jackson)
+        .withGrpc
         .withTests()
         .withE2e()